requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0

# PDF processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# NLP
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
psutil>=5.9.0

# Interface
streamlit>=1.25.0
flask>=2.3.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

